requirements.txt
# local package
-e .
langchain==0.1.7
langchain-openai==0.0.6
openai==1.12.0
pydantic==2.6.1
huggingface-hub
pytest
colorama
python-dotenv>=0.5.1
pyyaml
questionary
ruff
sentence-transformers
chromadb
typer
rich
pytest
pytest-mock
esprima
escodegen
langfuse

# external requirements
click
Sphinx


